Blog/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
django-js-asset==1.2.2
django-mptt==0.13.4
django-tinymce==3.4.0
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
